requirements.txt
streamlit
python-dotenv
groq
pyaudio
plotly